requirements.txt
grpcio
grpcio-tools
python-dotenv
pydantic
pymongo
protobuf==5.28.0